+++ internal/campaigns/archive_fetcher.go
@@ -12,6 +12,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/src-cli/internal/api"
 	"github.com/sourcegraph/src-cli/internal/campaigns/graphql"
@@ -67,6 +68,11 @@ func unzipToTempDir(ctx context.Context, zipFile, tempDir, tempFilePrefix string
 	if err != nil {
 		return "", err
 	}
+
+	if err := os.Chmod(volumeDir, 0777); err != nil {
+		return "", err
+	}
+
 	return volumeDir, unzip(zipFile, volumeDir)
 }
 
@@ -86,6 +92,10 @@ func fetchRepositoryArchive(ctx context.Context, client api.Client, repo *graphq
 		return fmt.Errorf("unable to fetch archive (HTTP %d from %s)", resp.StatusCode, req.URL.String())
 	}
 
+	// Unlike the mkdirAll() calls elsewhere in this file, this is only giving
+	// us a temporary place on the filesystem to keep the archive. Since it's
+	// never mounted into the containers being run, we can keep these
+	// directories 0700 without issue.
 	if err := os.MkdirAll(filepath.Dir(dest), 0700); err != nil {
 		return err
 	}
@@ -130,19 +140,29 @@ func unzip(zipFile, dest string) error {
 		}
 
 		if f.FileInfo().IsDir() {
-			if err := os.MkdirAll(fpath, os.ModePerm); err != nil {
+			if err := mkdirAll(dest, f.Name, 0777); err != nil {
 				return err
 			}
 			continue
 		}
 
-		if err := os.MkdirAll(filepath.Dir(fpath), os.ModePerm); err != nil {
+		if err := mkdirAll(dest, filepath.Dir(f.Name), 0777); err != nil {
 			return err
 		}
 
 		outFile, err := os.OpenFile(fpath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
 		if err != nil {
 			return err
+		}
+
+		// Since the container might not run as the same user, we need to ensure
+		// that the file is globally writable. If the execute bit is normally
+		// set on the zipped up file, let's ensure we propagate that to the
+		// group and other permission bits too.
+		if f.Mode()&0111 != 0 {
+			outFile.Chmod(0777)
+		} else {
+			outFile.Chmod(0666)
 		}
 
 		rc, err := f.Open()
@@ -165,4 +185,62 @@ func unzip(zipFile, dest string) error {
 	}
 
 	return nil
+}
+
+// Technically, this is a misnomer, since it might be a socket or block special,
+// but errPathExistsAsNonDir is just ugly for an internal type.
+type errPathExistsAsFile string
+
+var _ error = errPathExistsAsFile("")
+
+func (e errPathExistsAsFile) Error() string {
+	return fmt.Sprintf("path already exists, but not as a directory: %s", string(e))
+}
+
+// mkdirAll is essentially os.MkdirAll(filepath.Join(base, path), perm), but
+// applies the given permission regardless of the user's umask.
+func mkdirAll(base, path string, perm os.FileMode) error {
+	abs := filepath.Join(base, path)
+
+	// Create the directory if it doesn't exist.
+	st, err := os.Stat(abs)
+	if err != nil {
+		// It's expected that we'll get an error if the directory doesn't exist,
+		// so let's check that it's of the type we expect.
+		if !os.IsNotExist(err) {
+			return err
+		}
+
+		// Now we're clear to create the directory.
+		if err := os.MkdirAll(abs, perm); err != nil {
+			return err
+		}
+	} else if !st.IsDir() {
+		// The file/socket/whatever exists, but it's not a directory. That's
+		// definitely going to be an issue.
+		return errPathExistsAsFile(abs)
+	}
+
+	// If os.MkdirAll() was invoked earlier, then the permissions it set were
+	// subject to the umask. Let's walk the directories we may or may not have
+	// created and ensure their permissions look how we want.
+	return ensureAll(base, path, perm)
+}
+
+// ensureAll ensures that all directories under path have the expected
+// permissions.
+func ensureAll(base, path string, perm os.FileMode) error {
+	var errs *multierror.Error
+
+	// In plain English: for each directory in the path parameter, we should
+	// chmod that path to the permissions that are expected.
+	acc := []string{base}
+	for _, element := range strings.Split(path, string(os.PathSeparator)) {
+		acc = append(acc, element)
+		if err := os.Chmod(filepath.Join(acc...), perm); err != nil {
+			errs = multierror.Append(errs, err)
+		}
+	}
+
+	return errs.ErrorOrNil()
 }