+++ internal/campaigns/run_steps.go
@@ -112,6 +112,19 @@ func runSteps(ctx context.Context, wc *WorkspaceCreator, repo *graphql.Repositor
 			return nil, errors.Wrap(err, "closing temporary file")
 		}
 
+		// This file needs to be readable within the container regardless of the
+		// user the container is running as, so we'll set the appropriate group
+		// and other bits to make it so.
+		//
+		// A fun note: although os.File exposes a Chmod() method, we can't
+		// unconditionally use it because Windows cannot change the attributes
+		// of an open file. Rather than going to the trouble of having
+		// conditionally compiled files here, instead we'll just wait until the
+		// file is closed to twiddle the permission bits. Which is now!
+		if err := os.Chmod(runScriptFile.Name(), 0644); err != nil {
+			return nil, errors.Wrap(err, "setting permissions on the temporary file")
+		}
+
 		// Parse and render the step.Files.
 		files, err := renderStepFiles(step.Files, &stepContext)
 		if err != nil {